requirementes2.txt
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.4.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
Brlapi==0.6.6
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
command-not-found==0.3
cryptography==2.9
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0.5
django-admin==2.0.1
django-allauth==0.41.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-haystack==2.8.1
django-six==1.0.4
djangorestframework==3.11.0
entrypoints==0.3
gpg==1.10.0
gunicorn==20.0.4
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
notebook==6.0.3
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.0.0
psycopg2==2.8.5
ptyprocess==0.6.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.5.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-magic==0.4.16
python3-openid==3.1.0
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
screen==1.0.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.25.8
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.0.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==4.3.2
